f/search-form linked up with bootstrap CSS

+++ server/db/models/order.js
@@ -16,34 +16,40 @@ const Order = db.define('order', {
     defaultValue: 'cart'
   },
   cartTotal: {
-    type: Sequelize.DECIMAL(10, 2),
+    type: Sequelize.INTEGER,
     defaultValue: 0
   },
   orderTotal: {
-    type: Sequelize.DECIMAL(10, 2)
+    type: Sequelize.INTEGER
   }
 })
 
 Order.findOrCreateCartByUserId = async function(userId) {
   try {
     let cart = await this.findOne({where: {userId, status: 'cart'}})
-    if (!cart) cart = await this.create({
-      status: 'cart',
-      userId
-    }) 
+    if (!cart)
+      cart = await this.create({
+        status: 'cart',
+        userId
+      })
     return cart
-  } catch (e) { console.error(e) }
+  } catch (e) {
+    console.error(e)
+  }
 }
 
 const forceOneCart = async order => {
   if (order.status === 'cart') {
-    try { 
-      const { userId } = order.dataValues
+    try {
+      const {userId} = order.dataValues
       const existingCart = await Order.findOne({
         where: {userId, status: 'cart'}
       })
-      if (existingCart) throw new Error(`User (id: ${userId}) can only have one cart.`)
-    } catch (e) { console.error(e) }
+      if (existingCart)
+        throw new Error(`User (id: ${userId}) can only have one cart.`)
+    } catch (e) {
+      console.error(e)
+    }
   }
 }
 
@@ -56,7 +62,9 @@ const setCartTotal = async order => {
     })
     const total = items.reduce((a, b) => a + b.price, 0)
     order.cartTotal = total
-  } catch (e) { console.error(e) }
+  } catch (e) {
+    console.error(e)
+  }
 }
 
 Order.beforeCreate(forceOneCart)
